change(#73): Refactor that LaravelCms know how to bootstrap everything.
