--- resources/js/bootstrap/mixins.js
@@ -1,5 +1,0 @@
-import Vue from 'vue'
-import route from '../mixins/route';
-
-// global mixins
-Vue.mixin(route);
